rename PeripheralPuncher to RegisterRouter
